clauditor-kut.1: Move auth helpers to _providers/_auth.py (#144 US-001)
Create the _providers/ package and relocate the 9 auth-related symbols
(check_any_auth_available, check_api_key_only, announce_implicit_no_api_key,
_announced_implicit_no_api_key, _IMPLICIT_NO_API_KEY_ANNOUNCEMENT,
_AUTH_MISSING_TEMPLATE, _AUTH_MISSING_TEMPLATE_KEY_ONLY, _api_key_is_set,
_claude_cli_is_available) into _providers/_auth.py per DEC-005.

AnthropicAuthMissingError is defined exactly once at the package level
(_providers/__init__.py) and re-exported from both _providers and the
back-compat _anthropic shim, preserving the class-identity invariant
called out in Architecture Review's Security concern #1. A new
TestExceptionClassIdentity regression test pins this.

Tests that patch the announcement one-shot flag now target its canonical
home (clauditor._providers._auth._announced_implicit_no_api_key) instead
of the re-exported alias on _anthropic.
